examples/javascript/requirements.txt
Babel==2.6.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.1.0
Pygments==2.3.1
Sphinx==1.8.3
alabaster==0.7.12
certifi==2018.11.29
chardet==3.0.4
commonmark==0.8.1
docutils==0.14
future==0.17.1
git+https://github.com/codejamninja/recommonmark.git
html2text==2018.1.9
idna==2.8
imagesize==1.1.0
packaging==19.0
parsimonious==0.7.0
pyparsing==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
sphinx-js==2.3.1
sphinx-markdown-builder==0.2.0
sphinx-markdown-tables==0.0.9
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.1.0
typing==3.6.6
urllib3==1.24.1
